:package: move auth/ft_callback
