+++ src/nestjs/src/modules/auth/auth.module.ts
@@ -1,11 +1,14 @@
 import { Module } from "@nestjs/common";
+import { JwtModule } from "@nestjs/jwt";
+import { PassportModule } from "@nestjs/passport";
+import { APP_GUARD } from "@nestjs/core";
+
 import { AuthController } from "./auth.controller";
 import { AuthService } from "./auth.service";
-import { JwtModule } from "@nestjs/jwt";
 import { Api42Module } from "../api42/api42.module";
 import { DbModule } from "../database/database.module";
-import { PassportModule } from "@nestjs/passport";
 import { JwtStrategy } from "./jwt.strategy";
+import { JwtAuthGuard } from "./jwt-auth.guard";
 
 @Module({
 	imports: [
@@ -19,7 +22,14 @@ import { JwtStrategy } from "./jwt.strategy";
 		PassportModule.register({ defaultStrategy: "jwt" }),
 	],
 	controllers: [AuthController],
-	providers: [AuthService, JwtStrategy],
+	providers: [
+		AuthService,
+		JwtStrategy,
+		{
+			provide: APP_GUARD,
+			useClass: JwtAuthGuard,
+		}
+	],
 	exports: [AuthService, JwtModule],
 })
 export class AuthModule {}
